settled on which main.py is correct.
